requirements.txt
Django==1.10.5
psycopg2==2.7.3.2
Pillow==4.1.1
django-imagekit==4.0.1
django-ses==0.8.2
polyline==1.3.2
requests==2.9.1
gpxpy==1.1.2
django-jenkins==0.110.0